app/agent/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.7

# Vector database support
pgvector==0.2.4

# AI and OpenAI
openai==1.3.5
swarm-ai==1.0.0

# HTTP client
httpx==0.25.2

# Data validation and serialization
pydantic==2.5.0

# Environment and configuration
python-dotenv==1.0.0

# Logging and monitoring
structlog==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
